Update spatial_ar1_data.R
So that we can generate new data for each iteration, I've commented out the seed and set it up so that it can be called as part of another script (where the seed would be set) by using source(). From there we can then analyze our data, with that code to follow.